Refactored to put all data processing code into a helper class.
